@@ -14,7 +14,7 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@ark-ui/react": "^1.2.1",
+    "@ark-ui/react": "^2.0.1",
     "@hookform/resolvers": "^3.3.4",
     "@radix-ui/react-aspect-ratio": "^1.0.3",
     "@radix-ui/react-collapsible": "^1.0.3",
@@ -64,7 +64,7 @@
     "tailwind-plugin-variants": "^1.0.2",
     "tailwindcss": "^3.4.0",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "^5.1.6",
+    "typescript": "^5.3.3",
     "vite": "^5.0.0",
     "vite-tsconfig-paths": "^4.2.1"
   },
