@@ -1,11 +1,10 @@
 import { Center } from "~/components/ui/stack";
-import { Accordions } from "~/demo/accordion.stories";
-import { Collapsibles } from "~/demo/collapsible.stories";
+import { DatePickers } from "~/demo/date-picker.stories";
 
 export default function Wip() {
   return (
     <Center className="flex-1">
-      <Accordions />
+      <DatePickers />
     </Center>
   );
 }
